refactor(web): converge Step 3 on the two facts it already had
The step had grown a second readiness state machine and a Bot × tools
copy matrix on top of it, so three surfaces each explained the same state
and kept disagreeing. That was over-built for what this is: the member
chooses Feishu, and the page calls the setup Task that already exists.

Deleted: `OpenTagToolsPrep`, `OpenTagFeishuStepState`, `OpenTagBotState`,
`resolveOpenTagFeishuStep`, `resolveFeishuBotState`, `feishuStepCopy`, the
shell's runtime heading override, the readiness panel's narration, and the
table-driven combination tests that pinned those derivations.

What is left is what the four invariants need. Step 3 keeps one neutral
heading. The two rows read their own halves straight from the binding and
`binding.cli`. Recovery is four plain booleans on the page — a failed
setup call, a failed Bot, no Computer, or a long wait — and a finished
handoff clears it. `Try again` shows only where a Computer could still
finish; `Finish later` always does.

The server helper moves back inline next to the route that was already
holding this prompt, so the endpoint keeps its create-or-reuse and its
explicit retry without a module of its own.

Regression coverage is now observable behavior only: the call is not made
before the member chooses Feishu, not for a teammate's Agent, and not for
an Agent with no Computer; it is made once per Agent and Computer; a
retry carries the retry intent; first use and completion need both halves;
and each failure offers the way out at once.
