+++ packages/web/src/pages/opentag/__tests__/flow.test.ts
@@ -1,17 +1,7 @@
 import type { FeishuBotBinding } from "@first-tree/shared";
 import { describe, expect, it } from "vitest";
 import { shouldEnterOnboarding } from "../../onboarding/steps.js";
-import { feishuStepCopy } from "../copy.js";
-import {
-  classifyOpenTagAgent,
-  isFeishuHandoffUsable,
-  type OpenTagAgentRead,
-  type OpenTagBotState,
-  type OpenTagToolsPrep,
-  resolveFeishuBotState,
-  resolveOpenTagFeishuStep,
-  resolveOpenTagStep,
-} from "../flow.js";
+import { classifyOpenTagAgent, isFeishuHandoffUsable, type OpenTagAgentRead, resolveOpenTagStep } from "../flow.js";
 
 const ORG = "org-1";
 
@@ -208,99 +198,6 @@ describe("why the readiness refresh has to be authoritative", () => {
   });
 });
 
-describe("what the Feishu step is showing and what it offers", () => {
-  const BINDING: FeishuBotBinding = {
-    id: "binding-1",
-    agentId: "agent-1",
-    appId: "cli_1",
-    botOpenId: null,
-    botName: null,
-    botAvatarUrl: null,
-    tenantKey: null,
-    status: "active",
-    connectionStatus: "connected",
-    grantedScopes: [],
-    registrationUrl: null,
-    registrationExpiresAt: null,
-    lastConnectedAt: null,
-    lastEventAt: null,
-    lastErrorCode: null,
-    lastErrorMessage: null,
-    cli: { state: "missing", version: null, clientId: "client-1" },
-  };
-  const READY_CLI = { state: "ready" as const, version: "1.4.0", clientId: "client-1" };
-  const waiting = { binding: BINDING, callFailed: false, slow: false };
-
-  it("prepares nothing for a member who has not chosen Feishu", () => {
-    // No Bot means no commitment, and a machine check nobody asked for is
-    // exactly the install concept this flow exists to keep off the member.
-    expect(resolveOpenTagFeishuStep({ binding: null, callFailed: false, slow: false })).toEqual({
-      tools: { state: "idle" },
-      recovery: "none",
-      canRetryTools: false,
-    });
-  });
-
-  it("asks nothing of the member while the wait is ordinary", () => {
-    expect(resolveOpenTagFeishuStep(waiting)).toEqual({
-      tools: { state: "preparing" },
-      recovery: "none",
-      canRetryTools: false,
-    });
-  });
-
-  it("offers a way forward once waiting has stopped being reasonable", () => {
-    expect(resolveOpenTagFeishuStep({ ...waiting, slow: true })).toEqual({
-      tools: { state: "recoverable", reason: "slow" },
-      recovery: "offered",
-      canRetryTools: true,
-    });
-    expect(resolveOpenTagFeishuStep({ ...waiting, callFailed: true })).toEqual({
-      tools: { state: "recoverable", reason: "failed" },
-      recovery: "offered",
-      canRetryTools: true,
-    });
-  });
-
-  it("says the request failed rather than that it is taking a while", () => {
-    // A request that never landed is established, not suspected: the member is
-    // not waiting for anything, and a timeout that measures nothing would tell
-    // them to keep waiting for work that was never started.
-    expect(resolveOpenTagFeishuStep({ ...waiting, callFailed: true, slow: true }).tools).toEqual({
-      state: "recoverable",
-      reason: "failed",
-    });
-  });
-
-  it("reports the capability, not the history that produced it", () => {
-    // The Agent may well have failed once and recovered on its own. What the
-    // member needs is whether their Agent can work in Feishu now.
-    const binding = { ...BINDING, cli: READY_CLI };
-    expect(resolveOpenTagFeishuStep({ binding, callFailed: true, slow: true }).tools).toEqual({ state: "ready" });
-  });
-
-  it("strands nobody when the Bot is the half that never lands", () => {
-    // A Computer that was ready all along sends no request, so timing the
-    // request would leave exactly this member with no way out: the Bot is stuck,
-    // and the page offers nothing. Retrying is not among the ways forward —
-    // the automatic preparation has nothing left to do.
-    const stuck = { ...BINDING, status: "provisioning" as const, connectionStatus: "disconnected" as const };
-    const state = resolveOpenTagFeishuStep({ binding: { ...stuck, cli: READY_CLI }, callFailed: false, slow: true });
-    expect(state).toEqual({ tools: { state: "ready" }, recovery: "offered", canRetryTools: false });
-  });
-
-  it("offers nothing once the handoff actually works", () => {
-    // A long wait that ended well is not a problem to recover from — and
-    // neither is one that started with a failed request and then converged.
-    // The Agent may well have recovered on its own; offering a way out beside
-    // an invitation to go and use the Bot would contradict it.
-    const usable = { ...BINDING, cli: READY_CLI };
-    expect(resolveOpenTagFeishuStep({ binding: usable, callFailed: false, slow: true }).recovery).toBe("none");
-    expect(resolveOpenTagFeishuStep({ binding: usable, callFailed: true, slow: false }).recovery).toBe("none");
-    expect(resolveOpenTagFeishuStep({ binding: usable, callFailed: true, slow: true }).recovery).toBe("none");
-  });
-});
-
 describe("when the Feishu handoff is genuinely usable", () => {
   const usable: FeishuBotBinding = {
     id: "binding-1",
@@ -341,163 +238,5 @@ describe("when the Feishu handoff is genuinely usable", () => {
   it("refuses a ready Computer with no reachable Bot", () => {
     expect(isFeishuHandoffUsable({ ...usable, connectionStatus: "connecting" })).toBe(false);
     expect(isFeishuHandoffUsable({ ...usable, status: "provisioning" })).toBe(false);
-  });
-});
-
-describe("what the Feishu step says about itself", () => {
-  const TOOLS: OpenTagToolsPrep[] = [
-    { state: "preparing" },
-    { state: "ready" },
-    { state: "unavailable" },
-    { state: "recoverable", reason: "slow" },
-    { state: "recoverable", reason: "failed" },
-  ];
-  const BOTS: OpenTagBotState[] = ["waiting", "reachable", "failed"];
-
-  /** Language that promises work is under way on the Agent's Computer. */
-  const CLAIMS_WORK_UNDERWAY = [
-    "still preparing",
-    "keep checking this Computer",
-    "Finishing Agent tools",
-    "Preparing this Computer",
-  ];
-  /** Language that frames the Bot as merely unconfirmed. */
-  const CLAIMS_BOT_MERELY_WAITING = ["hasn't confirmed", "Waiting for Feishu", "both parts finish"];
-
-  function said(bot: OpenTagBotState, tools: OpenTagToolsPrep): string {
-    const { heading, panel } = feishuStepCopy(bot, tools);
-    return `${heading?.why ?? ""} ${heading?.lead ?? ""} ${panel?.title ?? ""} ${panel?.lead ?? ""}`;
-  }
-
-  it("never claims the Computer is busy when it is done, missing, or was never asked", () => {
-    // Three surfaces report the same two facts — the page heading, the panel
-    // header, and the rows. Deriving them separately is exactly how they came
-    // to contradict each other, so this pins every combination at once.
-    for (const tools of TOOLS) {
-      if (tools.state === "preparing" || (tools.state === "recoverable" && tools.reason === "slow")) continue;
-      for (const bot of BOTS) {
-        for (const claim of CLAIMS_WORK_UNDERWAY) {
-          expect(`${tools.state}/${bot}: ${said(bot, tools)}`).not.toContain(claim);
-        }
-      }
-    }
-  });
-
-  it("never frames a failed Bot as one the member simply has not confirmed", () => {
-    // The Bot column and its readiness row report the error; a panel promising
-    // that Feishu is about to confirm it says the opposite of both.
-    for (const tools of TOOLS) {
-      for (const claim of CLAIMS_BOT_MERELY_WAITING) {
-        expect(`${tools.state}: ${said("failed", tools)}`).not.toContain(claim);
-      }
-    }
-  });
-
-  it("says nothing at all once both halves are done", () => {
-    const done = feishuStepCopy("reachable", { state: "ready" });
-    expect(done.heading).toBeNull();
-    expect(done.panel).toBeNull();
-  });
-
-  it("keeps the standing heading while waiting is still ordinary", () => {
-    expect(feishuStepCopy("reachable", { state: "preparing" }).heading).toBeNull();
-    expect(feishuStepCopy("waiting", { state: "preparing" }).heading).toBeNull();
-  });
-
-  it("reports a request that never landed as failed, not as slow", () => {
-    const { heading, panel } = feishuStepCopy("reachable", { state: "recoverable", reason: "failed" });
-    expect(heading?.why).toContain("couldn't start");
-    expect(panel?.title).toContain("didn't start");
-  });
-
-  it("names the missing Computer rather than describing preparation", () => {
-    const { heading, panel } = feishuStepCopy("reachable", { state: "unavailable" });
-    expect(heading?.why).toContain("no Computer");
-    expect(panel?.title).toContain("can't be prepared");
-  });
-
-  it("names the Bot failure even when the Computer is ready", () => {
-    const { heading, panel } = feishuStepCopy("failed", { state: "ready" });
-    expect(heading?.why).toContain("needs attention");
-    expect(panel?.lead).toContain("The Bot is the part that needs you.");
-  });
-
-  it("keeps the approved copy for the state the contract pictures", () => {
-    const { heading, panel } = feishuStepCopy("reachable", { state: "recoverable", reason: "slow" });
-    expect(heading?.why).toBe("One last part is taking longer.");
-    expect(heading?.lead).toBe(
-      "Your Feishu Bot is connected. Your Agent's Computer is still preparing the tools it needs to reply.",
-    );
-    expect(panel?.title).toBe("Finishing Agent tools…");
-  });
-});
-
-describe("telling the Bot's three answers apart", () => {
-  function bot(overrides: Partial<FeishuBotBinding>): FeishuBotBinding {
-    return {
-      id: "b",
-      agentId: "a",
-      appId: null,
-      botOpenId: null,
-      botName: null,
-      botAvatarUrl: null,
-      tenantKey: null,
-      status: "provisioning",
-      connectionStatus: "disconnected",
-      grantedScopes: [],
-      registrationUrl: null,
-      registrationExpiresAt: null,
-      lastConnectedAt: null,
-      lastEventAt: null,
-      lastErrorCode: null,
-      lastErrorMessage: null,
-      cli: { state: "ready", version: null, clientId: "c" },
-      ...overrides,
-    };
-  }
-
-  it("separates a failure from a confirmation that has not happened yet", () => {
-    expect(resolveFeishuBotState(bot({ status: "error" }))).toBe("failed");
-    expect(resolveFeishuBotState(bot({ status: "active", connectionStatus: "error" }))).toBe("failed");
-    expect(resolveFeishuBotState(bot({}))).toBe("waiting");
-    expect(resolveFeishuBotState(bot({ status: "active", connectionStatus: "connecting" }))).toBe("waiting");
-    expect(resolveFeishuBotState(bot({ status: "active", connectionStatus: "connected" }))).toBe("reachable");
-    expect(resolveFeishuBotState(null)).toBe("waiting");
-  });
-});
-
-describe("an Agent with no Computer to prepare", () => {
-  const OFFLINE: FeishuBotBinding = {
-    id: "binding-1",
-    agentId: "agent-1",
-    appId: "cli_1",
-    botOpenId: null,
-    botName: null,
-    botAvatarUrl: null,
-    tenantKey: null,
-    status: "active",
-    connectionStatus: "connected",
-    grantedScopes: [],
-    registrationUrl: null,
-    registrationExpiresAt: null,
-    lastConnectedAt: null,
-    lastEventAt: null,
-    lastErrorCode: null,
-    lastErrorMessage: null,
-    cli: { state: "offline", version: null, clientId: null },
-  };
-
-  it("does not wait out a clock for something that cannot start", () => {
-    // No Computer is established, like a failed request — not suspected, like a
-    // slow one. Sitting on the 90-second timer would leave the member watching
-    // preparation that can never begin.
-    const state = resolveOpenTagFeishuStep({ binding: OFFLINE, callFailed: false, slow: false });
-    expect(state.tools).toEqual({ state: "unavailable" });
-    expect(state.recovery).toBe("offered");
-  });
-
-  it("does not offer to retry what has nowhere to run", () => {
-    const state = resolveOpenTagFeishuStep({ binding: OFFLINE, callFailed: false, slow: true });
-    expect(state.canRetryTools).toBe(false);
   });
 });
