+++ packages/web/src/pages/opentag/copy.ts
@@ -1,4 +1,4 @@
-import type { OpenTagBotState, OpenTagStepId, OpenTagToolsPrep } from "./flow.js";
+import type { OpenTagStepId } from "./flow.js";
 
 /**
  * Member-facing strings for the `/opentag` entry, kept out of the components
@@ -38,7 +38,7 @@ export const OPENTAG_STEP_COPY: Record<OpenTagStepId, { why: string; lead: strin
   },
   "connect-feishu": {
     why: "Connect your Agent to Feishu.",
-    lead: "Confirm the Bot in Feishu while First Tree prepares everything your Agent needs in the background.",
+    lead: "Confirm the Bot in Feishu. Both parts have to be ready before your Agent can work there.",
   },
   // Says only what the Task itself establishes. The shell renders this heading
   // for the whole step, including while the completion stamp is still in flight
@@ -52,127 +52,13 @@ export const OPENTAG_STEP_COPY: Record<OpenTagStepId, { why: string; lead: strin
 
 /** Member-facing strings for the two things Step 3 is waiting on. */
 export const OPENTAG_FEISHU_READINESS_COPY = {
-  panelPreparingTitle: "Preparing First Tree for Feishu…",
-  panelPreparingLead: "You can stay on this page while both parts finish.",
-  panelDelayedTitle: "Finishing Agent tools…",
-  panelDelayedLead: "First Tree will keep checking this Computer.",
   botLabel: "Feishu Bot",
   toolsLabel: "Agent tools",
   recoveryTitle: "You are not stuck here.",
   recoveryLead: "Try the automatic setup again, or finish later and repair it from this Agent's settings.",
-  // When the Computer is already done, the automatic setup has nothing left to
-  // retry — the outstanding half is the member's own confirmation in Feishu.
+  // When there is nothing left for the automatic setup to do, leaving is the
+  // only honest offer.
   recoveryLeadFinishOnly: "Finish later and pick this up from this Agent's settings whenever you're ready.",
   tryAgain: "Try again",
   finishLater: "Finish later",
 } as const;
-
-/**
- * Everything the Feishu step says about itself, derived in one place from one
- * state.
- *
- * The step reports the same two facts three times over — a page heading, a
- * panel header, and the readiness rows — and three separate derivations of the
- * same thing is how they came to contradict each other. Deriving all of them
- * from the resolved state is what keeps "the automatic setup didn't start"
- * from appearing under "First Tree will keep checking this Computer".
- *
- * `heading` is null while the standing step heading is still the right
- * question, and `panel` is null once both halves are done and there is nothing
- * left to narrate.
- */
-export function feishuStepCopy(
-  bot: OpenTagBotState,
-  tools: OpenTagToolsPrep,
-): {
-  heading: { why: string; lead: string } | null;
-  panel: { title: string; lead: string } | null;
-} {
-  const copy = OPENTAG_FEISHU_READINESS_COPY;
-  const bothWaiting = { title: copy.panelPreparingTitle, lead: copy.panelPreparingLead };
-  const botReachable = bot === "reachable";
-
-  // A Bot that failed outranks whatever the Computer is doing: it is the more
-  // severe outstanding fact, and describing it as a wait — or as something that
-  // will finish on its own — contradicts the error shown right beside this.
-  if (bot === "failed") {
-    return {
-      heading: {
-        why: "Feishu setup needs attention.",
-        lead: "Setting up the Bot didn't finish, so your Agent can't be reached in Feishu yet.",
-      },
-      panel: {
-        title: "Feishu Bot needs attention.",
-        lead:
-          tools.state === "ready"
-            ? "Your Agent is ready. The Bot is the part that needs you."
-            : "Setting up the Bot didn't finish.",
-      },
-    };
-  }
-
-  switch (tools.state) {
-    case "unavailable":
-      return {
-        heading: {
-          why: "This Agent has no Computer yet.",
-          lead: "Feishu work runs on the Agent's own Computer. You can give it one from this Agent's settings.",
-        },
-        panel: {
-          title: "Agent tools can't be prepared yet.",
-          lead: "There is no Computer for this Agent to prepare them on.",
-        },
-      };
-    case "ready":
-      // The Computer is done. Anything that named it as outstanding would sit
-      // directly above a row reporting it ready.
-      return botReachable
-        ? { heading: null, panel: null }
-        : {
-            heading: {
-              why: "This is taking longer than usual.",
-              lead: "Your Agent is ready. Feishu hasn't confirmed the Bot yet.",
-            },
-            panel: { title: "Waiting for Feishu…", lead: "Your Agent is ready. Confirming the Bot is the last part." },
-          };
-    case "recoverable":
-      if (tools.reason === "failed") {
-        // A request that never landed is not a duration, and nothing is being
-        // checked in the background — so neither the heading nor the panel may
-        // describe work in progress.
-        return {
-          heading: {
-            why: "We couldn't start the setup.",
-            lead: botReachable
-              ? "Your Feishu Bot is connected. First Tree couldn't start preparing your Agent's Computer."
-              : "First Tree couldn't start preparing your Agent's Computer.",
-          },
-          panel: {
-            title: "Agent tools didn't start.",
-            lead: "Trying again asks your Agent to prepare this Computer.",
-          },
-        };
-      }
-      return botReachable
-        ? {
-            heading: {
-              why: "One last part is taking longer.",
-              lead: "Your Feishu Bot is connected. Your Agent's Computer is still preparing the tools it needs to reply.",
-            },
-            panel: { title: copy.panelDelayedTitle, lead: copy.panelDelayedLead },
-          }
-        : {
-            heading: {
-              why: "This is taking longer than usual.",
-              lead: "Your Agent's Computer is still preparing the tools it needs to reply in Feishu.",
-            },
-            panel: bothWaiting,
-          };
-    default:
-      // Ordinary waiting, on one half or both.
-      return {
-        heading: null,
-        panel: botReachable ? { title: copy.panelDelayedTitle, lead: copy.panelDelayedLead } : bothWaiting,
-      };
-  }
-}
